change float to double

+++ measure/plot.h
@@ -1,1 +1,1 @@
-void plot(float *x_value, int *y_value, int count, char *x_label, char *y_label, char *title, char *output_filename);
+void plot(double *x_value, int *y_value, int count, char *x_label, char *y_label, char *title, char *output_filename);
